cuda: simplify matching model
Using a match on name only is actually more robust, because
if there is an event that has the same name but different type,
it will result in a compile time error. If more strict match
is used, events that have the same parame name and don't match
would be silently ignored.